Reorganizing the net setup
