+++ megamek/src/megamek/client/Client.java
@@ -35,10 +35,13 @@ import megamek.common.enums.GamePhase;
 import megamek.common.event.*;
 import megamek.common.force.Force;
 import megamek.common.force.Forces;
-import megamek.common.net.*;
+import megamek.common.net.connections.AbstractConnection;
 import megamek.common.net.enums.PacketCommand;
 import megamek.common.net.events.DisconnectedEvent;
 import megamek.common.net.events.PacketReceivedEvent;
+import megamek.common.net.factories.ConnectionFactory;
+import megamek.common.net.listeners.ConnectionListener;
+import megamek.common.net.packets.Packet;
 import megamek.common.options.GameOptions;
 import megamek.common.options.IBasicOption;
 import megamek.common.preference.PreferenceManager;
