+++ megamek/src/megamek/common/net/connections/AbstractConnection.java
@@ -13,15 +13,17 @@
 * FOR A PARTICULAR PURPOSE. See the GNU General Public License for more
 * details.
 */
-package megamek.common.net;
+package megamek.common.net.connections;
 
+import megamek.common.net.listeners.ConnectionListener;
+import megamek.common.net.packets.Packet;
 import megamek.common.net.enums.PacketCommand;
 import megamek.common.net.events.AbstractConnectionEvent;
 import megamek.common.net.events.ConnectedEvent;
 import megamek.common.net.events.DisconnectedEvent;
 import megamek.common.net.events.PacketReceivedEvent;
-import megamek.common.net.marshall.PacketMarshaller;
-import megamek.common.net.marshall.PacketMarshallerFactory;
+import megamek.common.net.marshalling.PacketMarshaller;
+import megamek.common.net.marshalling.PacketMarshallerFactory;
 import org.apache.logging.log4j.LogManager;
 
 import java.io.*;
