+++ megamek/src/megamek/common/net/enums/ConnectionEventType.java
@@ -18,7 +18,7 @@
  */
 package megamek.common.net.enums;
 
-import megamek.common.net.ConnectionListener;
+import megamek.common.net.listeners.ConnectionListener;
 import megamek.common.net.events.ConnectedEvent;
 import megamek.common.net.events.AbstractConnectionEvent;
 import megamek.common.net.events.DisconnectedEvent;
