+++ megamek/src/megamek/server/Server.java
@@ -38,10 +38,10 @@ import megamek.common.event.GameVictoryEvent;
 import megamek.common.force.Force;
 import megamek.common.force.Forces;
 import megamek.common.icons.Camouflage;
-import megamek.common.net.AbstractConnection;
-import megamek.common.net.ConnectionFactory;
-import megamek.common.net.ConnectionListener;
-import megamek.common.net.Packet;
+import megamek.common.net.connections.AbstractConnection;
+import megamek.common.net.factories.ConnectionFactory;
+import megamek.common.net.listeners.ConnectionListener;
+import megamek.common.net.packets.Packet;
 import megamek.common.net.enums.PacketCommand;
 import megamek.common.net.events.DisconnectedEvent;
 import megamek.common.net.events.PacketReceivedEvent;
