+++ megamek/src/megamek/server/ServerLobbyHelper.java
@@ -24,7 +24,7 @@ import megamek.common.Game;
 import megamek.common.Player;
 import megamek.common.force.Force;
 import megamek.common.force.Forces;
-import megamek.common.net.Packet;
+import megamek.common.net.packets.Packet;
 import megamek.common.net.enums.PacketCommand;
 import megamek.common.options.OptionsConstants;
 import org.apache.logging.log4j.LogManager;
